Refactor: shift responsibility
